# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShareHub API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/resources/{id}/comments:
    get:
      summary: 获取资源评论树
      responses:
        '200':
          description: 成功
    post:
      summary: 发表评论
      description: 需要当前用户身份；评论内容为空时返回 `400 COMMENT_CONTENT_REQUIRED`；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '400':
          description: 评论内容为空（错误码 `COMMENT_CONTENT_REQUIRED`）
        '404':
          description: 资源不存在（错误码 `RESOURCE_NOT_FOUND`）
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
  /api/comments/{id}/reply:
    post:
      summary: 回复评论
      description: 需要当前用户身份；评论内容为空时返回 `400 COMMENT_CONTENT_REQUIRED`；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '400':
          description: 评论内容为空（错误码 `COMMENT_CONTENT_REQUIRED`）
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
        '404':
          description: 父评论不存在（错误码 `COMMENT_NOT_FOUND`）
  /api/resources/{id}/favorite:
    post:
      summary: 收藏资料
      description: 需要当前用户身份；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '404':
          description: 资源不存在（错误码 `RESOURCE_NOT_FOUND`）
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
    delete:
      summary: 取消收藏资料
      description: 需要当前用户身份；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '404':
          description: 资源不存在（错误码 `RESOURCE_NOT_FOUND`）
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
  /api/resources/{id}/like:
    post:
      summary: 点赞资料
      description: 需要当前用户身份；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '404':
          description: 资源不存在（错误码 `RESOURCE_NOT_FOUND`）
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
    delete:
      summary: 取消点赞资料
      description: 需要当前用户身份；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '404':
          description: 资源不存在（错误码 `RESOURCE_NOT_FOUND`）
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
...
  /api/reports:
    post:
      summary: 举报内容
      description: 需要当前用户身份；支持 `RESOURCE` 与 `NOTE` 两种目标，资源举报传 `resourceId`，笔记举报传 `targetType=NOTE` + `noteId`；`reason` 未传、为空字符串或仅空白时回落为 `无`；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      responses:
        '200':
          description: 成功
        '404':
          description: 目标不存在（错误码 `RESOURCE_NOT_FOUND` / `NOTE_NOT_FOUND`）
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
...